script/requirements.txt
astroid==2.1.0
autopep8==1.4.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.4.0
pylint==2.2.2
six==1.12.0
typed-ast==1.1.0
wrapt==1.10.11
